Testing bringing out the kms client helper function.
